Separate Settings from UI, provide LiveData
